+++ src/bundle/Editor.ts
@@ -104,7 +104,6 @@ export interface EditorInt
 
 type SetEditorModeOptions = Pick<ChangeEditorModeOptions, 'emit'>;
 
-/** @internal */
 type ChangeEditorModeOptions = {
     mode: EditorMode;
     reason: 'error-boundary' | 'settings' | 'manually';
@@ -126,6 +125,11 @@ export type MarkupConfig = {
     languageData?: YfmLangOptions['languageData'];
     /** Part of autocompletion config. Can be used to style tooltip */
     autocompletionAddToOptions?: CreateCodemirrorParams['autocompletionAddToOptions'];
+};
+
+export type EscapeConfig = {
+    commonEscape?: RegExp;
+    startOfLineEscape?: RegExp;
 };
 
 export type EditorOptions = Pick<
@@ -154,12 +158,16 @@ export type EditorOptions = Pick<
      * You can use it to pre-process the value from the markup editor before it gets into the wysiwyg editor.
      */
     prepareRawMarkup?: (value: MarkupString) => MarkupString;
+    experimental_beforeEditorModeChange?: (
+        options: Pick<ChangeEditorModeOptions, 'mode' | 'reason'>,
+    ) => boolean | undefined;
     splitMode?: SplitMode;
     renderPreview?: RenderPreview;
     preset: EditorPreset;
     /** @deprecated Put extra extensions via MarkdownEditorMarkupConfig */
     extraMarkupExtensions?: CodemirrorExtension[];
     markupConfig?: MarkupConfig;
+    escapeConfig?: EscapeConfig;
 };
 
 /** @internal */
@@ -173,6 +181,7 @@ export class EditorImpl extends SafeEventEmitter<EventMapInt> implements EditorI
     #wysiwygEditor?: WysiwygEditor;
     #markupEditor?: MarkupEditor;
     #markupConfig: MarkupConfig;
+    #escapeConfig?: EscapeConfig;
 
     readonly #preset: EditorPreset;
     #allowHTML?: boolean;
@@ -183,6 +192,9 @@ export class EditorImpl extends SafeEventEmitter<EventMapInt> implements EditorI
     #fileUploadHandler?: FileUploadHandler;
     #needToSetDimensionsForUploadedImages: boolean;
     #prepareRawMarkup?: (value: MarkupString) => MarkupString;
+    #beforeEditorModeChange?: (
+        options: Pick<ChangeEditorModeOptions, 'mode' | 'reason'>,
+    ) => boolean | undefined;
 
     get _wysiwygView(): PMEditorView {
         // @ts-expect-error internal typing
@@ -274,6 +286,7 @@ export class EditorImpl extends SafeEventEmitter<EventMapInt> implements EditorI
                 allowHTML: this.#allowHTML,
                 linkify: this.#linkify,
                 linkifyTlds: this.#linkifyTlds,
+                escapeConfig: this.#escapeConfig,
                 onChange: () => this.emit('rerender-toolbar', null),
                 onDocChange: () => this.emit('change', null),
             });
@@ -343,6 +356,8 @@ export class EditorImpl extends SafeEventEmitter<EventMapInt> implements EditorI
             opts.needToSetDimensionsForUploadedImages,
         );
         this.#prepareRawMarkup = opts.prepareRawMarkup;
+        this.#escapeConfig = opts.escapeConfig;
+        this.#beforeEditorModeChange = opts.experimental_beforeEditorModeChange;
     }
 
     // ---> implements CodeEditor
@@ -379,8 +394,14 @@ export class EditorImpl extends SafeEventEmitter<EventMapInt> implements EditorI
 
     changeEditorMode({emit = true, ...opts}: ChangeEditorModeOptions): void {
         if (this.#editorMode === opts.mode) return;
+
+        if (this.#beforeEditorModeChange?.({mode: opts.mode, reason: opts.reason}) === false) {
+            return;
+        }
+
         this.currentMode = opts.mode;
         this.emit('rerender', null);
+
         if (emit) {
             this.emit('change-editor-mode', opts);
         }
@@ -486,6 +507,7 @@ export class EditorImpl extends SafeEventEmitter<EventMapInt> implements EditorI
     private shouldReplaceMarkupEditorValue(markupValue: string, wysiwygValue: string) {
         const serializedEditorMarkup = this.#wysiwygEditor?.serializer.serialize(
             this.#wysiwygEditor.parser.parse(markupValue),
+            this.#escapeConfig,
         );
         return serializedEditorMarkup?.trim() !== wysiwygValue.trim();
     }